requirements.txt
scikit-learn>=1.1.2
paddlepaddle-gpu==2.4.1.post112
paddleseg>=2.5
pymatting
scikit-image
numba
opencv-python
